测试(基础设施)：修复全量集成测试基线
为固定测试时间创建审计分区，避免 PostgreSQL 用例随当前月份漂移。

同步 Decimal 契约及当前 migration 的 schema 和字典基线。

+++ backend/crates/api/tests/m2_deferred_closeout_postgres.rs
@@ -473,7 +473,7 @@ async fn inspection_uses_actual_receipt_quantity_and_blocks_early_signature(pool
     assert_eq!(execution_evidence.1, approval_id);
     let putaway_task: (String, String, String, i64, Uuid) = sqlx::query_as(
         r#"
-        SELECT task_type_code, status, product_code, planned_qty, source_doc_id
+        SELECT task_type_code, status, product_code, planned_qty::BIGINT, source_doc_id
           FROM warehouse_tasks
          WHERE owner_id = $1
            AND source_doc_type = 'receiving_order'
